Rename source template to EmptyNonScrolling
To avoid renaming unintended names like Enumerable.Empty or EmptyScrollingGeneralCatchClause (in .sln.DotSettings)
