--- templates/ruleset-empty/osu.Game.Rulesets.Empty/EmptyInputManager.cs
@@ -1,29 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Collections.Generic;
-using System.ComponentModel;
-using osu.Framework.Input.Bindings;
-using osu.Game.Rulesets.UI;
-
-namespace osu.Game.Rulesets.Empty
-{
-    public class EmptyInputManager : RulesetInputManager<EmptyAction>
-    {
-        public IEnumerable<EmptyAction> PressedActions => KeyBindingContainer.PressedActions;
-
-        public EmptyInputManager(RulesetInfo ruleset)
-            : base(ruleset, 0, SimultaneousBindingMode.Unique)
-        {
-        }
-    }
-
-    public enum EmptyAction
-    {
-        [Description("Button 1")]
-        Button1,
-
-        [Description("Button 2")]
-        Button2,
-    }
-}
