--- templates/ruleset-empty/osu.Game.Rulesets.Empty/EmptyRuleset.cs
@@ -1,82 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Collections.Generic;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Input.Bindings;
-using osu.Game.Beatmaps;
-using osu.Game.Graphics;
-using osu.Game.Rulesets.Difficulty;
-using osu.Game.Rulesets.Mods;
-using osu.Game.Rulesets.Empty.Beatmaps;
-using osu.Game.Rulesets.Empty.Mods;
-using osu.Game.Rulesets.Empty.UI;
-using osu.Game.Rulesets.UI;
-using osuTK;
-using osuTK.Graphics;
-
-namespace osu.Game.Rulesets.Empty
-{
-    public class EmptyRuleset : Ruleset
-    {
-        public EmptyRuleset(RulesetInfo rulesetInfo = null)
-            : base(rulesetInfo)
-        {
-        }
-
-        public override string Description => "a very empty ruleset";
-
-        public override DrawableRuleset CreateDrawableRulesetWith(IWorkingBeatmap beatmap, IReadOnlyList<Mod> mods) => new DrawableEmptyRuleset(this, beatmap, mods);
-
-        public override IBeatmapConverter CreateBeatmapConverter(IBeatmap beatmap) => new EmptyBeatmapConverter(beatmap);
-
-        public override DifficultyCalculator CreateDifficultyCalculator(WorkingBeatmap beatmap) => new EmptyDifficultyCalculator(this, beatmap);
-
-        public override IEnumerable<Mod> GetModsFor(ModType type)
-        {
-            switch (type)
-            {
-                case ModType.Automation:
-                    return new[] { new EmptyModAutoplay() };
-
-                default:
-                    return new Mod[] { null };
-            }
-        }
-
-        public override string ShortName => "empty";
-
-        public override IEnumerable<KeyBinding> GetDefaultKeyBindings(int variant = 0) => new[]
-        {
-            new KeyBinding(InputKey.Z, EmptyAction.Button1),
-            new KeyBinding(InputKey.X, EmptyAction.Button2),
-        };
-
-        public override Drawable CreateIcon() => new Icon(ShortName[0]);
-
-        public class Icon : CompositeDrawable
-        {
-            public Icon(char c)
-            {
-                InternalChildren = new Drawable[]
-                {
-                    new Circle
-                    {
-                        Size = new Vector2(20),
-                        Colour = Color4.White,
-                    },
-                    new SpriteText
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Text = c.ToString(),
-                        Font = OsuFont.Default.With(size: 18)
-                    }
-                };
-            }
-        }
-    }
-}
