--- templates/ruleset-empty/osu.Game.Rulesets.Empty/UI/EmptyPlayfield.cs
@@ -1,22 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Game.Rulesets.UI;
-
-namespace osu.Game.Rulesets.Empty.UI
-{
-    [Cached]
-    public class EmptyPlayfield : Playfield
-    {
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            AddRangeInternal(new Drawable[]
-            {
-                HitObjectContainer,
-            });
-        }
-    }
-}
