+++ templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling/Replays/EmptyNonScrollingFramedReplayInputHandler.cs
@@ -1,0 +1,32 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Collections.Generic;
+using System.Linq;
+using osu.Framework.Input.StateChanges;
+using osu.Game.Replays;
+using osu.Game.Rulesets.Replays;
+
+namespace osu.Game.Rulesets.EmptyNonScrolling.Replays
+{
+    public class EmptyNonScrollingFramedReplayInputHandler : FramedReplayInputHandler<EmptyNonScrollingReplayFrame>
+    {
+        public EmptyNonScrollingFramedReplayInputHandler(Replay replay)
+            : base(replay)
+        {
+        }
+
+        protected override bool IsImportant(EmptyNonScrollingReplayFrame frame) => frame.Actions.Any();
+
+        public override List<IInput> GetPendingInputs()
+        {
+            return new List<IInput>
+            {
+                new ReplayState<EmptyNonScrollingAction>
+                {
+                    PressedActions = CurrentFrame?.Actions ?? new List<EmptyNonScrollingAction>(),
+                }
+            };
+        }
+    }
+}
