+++ templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling/osu.Game.Rulesets.EmptyNonScrolling.csproj
@@ -1,0 +1,20 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup Label="Project">
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyTitle>osu.Game.Rulesets.Sample</AssemblyTitle>
+    <OutputType>Library</OutputType>
+    <PlatformTarget>AnyCPU</PlatformTarget>
+    <RootNamespace>osu.Game.Rulesets.EmptyNonScrolling</RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\**" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="ppy.osu.Framework" Version="2019.905.0" />
+    <PackageReference Include="ppy.osu.Game" Version="2019.905.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Resources\Samples\Gameplay" />
+    <Folder Include="Resources\Textures" />
+  </ItemGroup>
+</Project>
